Implements result component that receives info from search and performs the query
