+++ src/app/pages/homepage/homepage.component.html
@@ -9,7 +9,7 @@
                 </p>
             </div>
             <div class="row">
-                <button type="button" class="btn btn-primary">Try it out!</button>
+                <a href="search"><button type="button" class="btn btn-primary">Try it out!</button></a>
             </div>
         </div>
     </div>
