+++ src/app/pages/searchpage/searchpage.component.html
@@ -110,29 +110,10 @@
 
     <div class="row justify-content-end">
       <div class="col-auto">
-        <button type="button" class="btn btn-query" (click)="queryData()">Query</button>
+        <button type="button" class="btn btn-query" (click)="sendData()">Query</button>
       </div>
       <div class="col-auto">
         <button type="button" class="btn btn-clear">Clear</button>
       </div>
-    </div>
-    <div *ngIf="dataIsNotEmpty">
-      <button id="btdown" class="btn" export-to-csv (click)="exportToCsv($event.target)">Export to CSV</button>
-      <table class="table table-striped">
-        <thead>
-          <tr>
-            <th scope="col">Date</th>
-            <th scope="col">Name</th>
-            <th scope="col">FOB Value</th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr *ngFor="let value of data.edges">
-            <td>{{value.node.date}}</td>
-            <td>{{value.node.ncm.ncmNamePt}}</td>
-            <td>{{value.node.fobValue}}</td>
-          </tr>
-        </tbody>
-      </table>
     </div>
   </div>
